+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+### Fixed
+- Both router and client don't validate RPC message types [\#34](https://github.com/angiolep/akka-wamp/issues/34)
+
     
 ## [0.9.0] _ 2016-09-29